dynamic speedlimit retrieval from DB using the location variable, all happens in speedviolationcheck.js function
