+++ api/routes/rfdata.js
@@ -2,7 +2,7 @@ const express = require('express')
 const mongoose = require('mongoose')
 const router = express.Router()
 
-const rf_v = require('../functions/rfdata_vio')
+const speedviolationCheck = require('../functions/speedviolationCheck')
 
 // Import the model
 const Rfdata = require('../models/Rfdata')
@@ -45,18 +45,17 @@ router.post('/rfdata', async (req, res, done) => {
         if (req.body.rfd_id === 'ID-1') {
             console.log('Middleware Fired.')
             const rf_tag = req.body.rf_tag
+            const location = req.body.location
             const arry = []
             for await (const doc of Rfdata.find({rf_tag: rf_tag,}).limit(2).sort('-eventTime')) {
                 let timestamp = new Date(doc.eventTime)
-                console.log(`${doc.rf_tag} --> ${timestamp}`)
+                // console.log(`${doc.rf_tag} --> ${timestamp.toLocaleTimeString()}`)
                 arry.push(timestamp)
                 // arry.push(timestamp.replace("AM","").replace("PM","").replace(" ",""))
                 // timestamp = new Date(doc.eventTime).toLocaleTimeString()
             }
-            console.log('---------------')
-            console.log(arry)
-            console.log('-------end--------')
-            rf_v(arry)
+            // array generated by query res is passed to module.
+            speedviolationCheck(location, arry)
         }
         
     } catch (err) {
